requirements.txt
python-frontmatter==1.1.0
httpx[http2]==0.28.1
tenacity==9.0.0
pyyaml==6.0.2
